Added microphone input.
Takes an input (responseTime) which is the amount of time the user has to perform the input. In this case it is making a noise. It reads the analog input pin for the microphone, and if the input is made in time, and it is loud enough, it will return true, otherwise it will return false.